paper fields added to NEO4J
