+++ ProjectX/joins.txt
@@ -1,3 +1,5 @@
-0	2	-0.1953125
-0	1	0.0
+1	2	-0.166752400549
+1	4	-0.0950360082305
+0	1	-0.0154749657064
+0	3	8.32667268469e-17
 END
