+++ ProjectX/levels.txt
@@ -3,6 +3,7 @@
 
 Initial: Level_0
 Going Downward: Level_-1
+Going Downward: Level_-2
 Going Upward: Level_+1
 Going Upward: Level_+2
 Going Upward: Level_+3: root
